requirements.txt
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
Automat==22.10.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.8
cssselect==1.2.0
defusedxml==0.7.1
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.4
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
iniconfig==2.0.0
itemadapter==0.9.0
itemloaders==1.3.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
jmespath==1.0.1
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==2.0.0
openai==1.57.3
orjson==3.10.6
packaging==24.1
pandas==2.2.2
parsel==1.9.1
playwright==1.48.0
pluggy==1.5.0
Protego==0.3.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic-extra-types==2.10.1
pydantic-settings==2.6.1
pydantic_core==2.20.1
PyDispatcher==2.0.7
pyee==12.0.0
Pygments==2.18.0
pyOpenSSL==24.1.0
PyPDF2==3.0.1
pytest==8.3.3
pytest-base-url==2.1.0
pytest-playwright==0.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
rich==13.7.1
Scrapy==2.11.2
service-identity==24.1.0
setuptools==70.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
text-unidecode==1.3
tldextract==5.1.2
tqdm==4.67.1
Twisted==24.3.0
twisted-iocpsupport==1.0.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
w3lib==2.2.1
watchfiles==0.22.0
websockets==12.0
zope.interface==6.4.post2
